feat: :fire: Introduce main script for the Master device `master.py`
This commit introduces the main script for the Master device, which serves as the central control unit in the IoT setup. The script is responsible for interacting with client devices over Modbus TCP. Key features of the script include:  - Establishing a Modbus TCP connection with a client device (BeagleBone Black). - Reading and updating data on the client device. - Configuring network connectivity on non-Docker MicroPython environments. - Requesting and setting data on the client device, including COILs (Control Relays) and HREGs (Holding Registers). - Resetting registers back to their default values on the client.  The script is designed to work in both Docker MicroPython and non-Docker MicroPython environments. It facilitates communication between the Master and client devices, making it a crucial component of the IoT system.